+++ ls_Toasts/systems/transmog.lua
@@ -3,6 +3,7 @@ local E, L, C = addonTable.E, addonTable.L, addonTable.C
 
 -- Lua
 local _G = getfenv(0)
+local t_wipe = _G.table.wipe
 
 -- Mine
 local function Toast_OnClick(self)
@@ -65,11 +66,34 @@ local function Toast_SetUp(event, sourceID, isAdded, attempt)
 	end
 end
 
+local pendingSourceIDs = {}
+local wipeTimer
+
+local function wiper()
+	t_wipe(pendingSourceIDs)
+end
+
+local function resetWipeTimer()
+	if not wipeTimer then
+		wipeTimer = C_Timer.NewTimer(5, wiper)
+	else
+		wipeTimer:Cancel()
+
+		wipeTimer = C_Timer.NewTimer(5, wiper)
+	end
+end
+
 local function TRANSMOG_COLLECTION_SOURCE_ADDED(sourceID)
 	-- don't show toasts for sources that aren't in player's wardrobe
 	local _, canCollect = C_TransmogCollection.PlayerCanCollectSource(sourceID)
 	if canCollect then
 		Toast_SetUp("TRANSMOG_COLLECTION_SOURCE_ADDED", sourceID, true, 1)
+	else
+		-- however, they may become available shortly after when
+		-- TRANSMOG_SOURCE_COLLECTABILITY_UPDATE fires
+		pendingSourceIDs[sourceID] = "TRANSMOG_COLLECTION_SOURCE_ADDED"
+
+		resetWipeTimer()
 	end
 end
 
@@ -81,6 +105,12 @@ local function TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED(sourceID)
 	local _, canCollect = C_TransmogCollection.PlayerCanCollectSource(sourceID)
 	if canCollect then
 		Toast_SetUp("TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED", sourceID, true, 1)
+	else
+		-- however, they may become available shortly after when
+		-- TRANSMOG_SOURCE_COLLECTABILITY_UPDATE fires
+		pendingSourceIDs[sourceID] = "TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED"
+
+		resetWipeTimer()
 	end
 end
 
@@ -92,11 +122,25 @@ local function TRANSMOG_COLLECTION_SOURCE_REMOVED(sourceID)
 	end
 end
 
+-- in some cases, for instance, quantum items, the source is marked as
+-- collectable only after the _ADDED events fire
+-- TRANSMOG_SOURCE_COLLECTABILITY_UPDATE usually fires a second or so later
+local function TRANSMOG_SOURCE_COLLECTABILITY_UPDATE(sourceID, isCollectable)
+	if isCollectable and pendingSourceIDs[sourceID] then
+		Toast_SetUp(pendingSourceIDs[sourceID], sourceID, true, 1)
+
+		pendingSourceIDs[sourceID] = nil
+
+		resetWipeTimer()
+	end
+end
+
 local function Enable()
 	if C.db.profile.types.transmog.enabled then
 		E:RegisterEvent("TRANSMOG_COLLECTION_SOURCE_ADDED", TRANSMOG_COLLECTION_SOURCE_ADDED)
 		E:RegisterEvent("TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED", TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED)
 		E:RegisterEvent("TRANSMOG_COLLECTION_SOURCE_REMOVED", TRANSMOG_COLLECTION_SOURCE_REMOVED)
+		E:RegisterEvent("TRANSMOG_SOURCE_COLLECTABILITY_UPDATE", TRANSMOG_SOURCE_COLLECTABILITY_UPDATE)
 	end
 end
 
@@ -104,6 +148,7 @@ local function Disable()
 	E:UnregisterEvent("TRANSMOG_COLLECTION_SOURCE_ADDED", TRANSMOG_COLLECTION_SOURCE_ADDED)
 	E:UnregisterEvent("TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED", TRANSMOG_COSMETIC_COLLECTION_SOURCE_ADDED)
 	E:UnregisterEvent("TRANSMOG_COLLECTION_SOURCE_REMOVED", TRANSMOG_COLLECTION_SOURCE_REMOVED)
+	E:UnregisterEvent("TRANSMOG_SOURCE_COLLECTABILITY_UPDATE", TRANSMOG_SOURCE_COLLECTABILITY_UPDATE)
 end
 
 local function Test()